Added BSc thesis
After getting the degree and sorting everything else out it was possible to upload the full thesis. Resulting data on which the thesis was based on has been uploaded too, as well as the results where a wrong approach was taken first (using the updated weights for error calculation in the backpropagation).